+++ std/jwt/mod.ts
@@ -1,3 +1,4 @@
 export { create } from "./create.ts";
 export { validate } from "./validate.ts";
 export { setExpiration } from "./_util.ts"
+export type { Algorithm, Header, Payload } from "./_util.ts";
